search page, structure: consolidate search components into SearchPage, rename some files
