+++ src/components/page-search/index.js
@@ -1,0 +1,1 @@
+export { default } from './page-search';
