+++ src/components/page-search/page-search.style.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Container = styled.article`
+  margin-top: 1.688rem; /* 27px */
+  text-align: center;
+`;
+
+export const Headline = styled.h1`
+  margin-top: 0
+`;
