--- src/hooks/use-search-form/index.js
@@ -1,1 +1,0 @@
-export { default } from './use-search-form';
