+++ src/hooks/use-subreddit-form/index.js
@@ -1,0 +1,1 @@
+export { default } from './use-subreddit-form';
